+++ handlers/whatsapp/whatsapp.go
@@ -103,6 +103,10 @@ type eventPayload struct {
 			MimeType string `json:"mime_type" validate:"required"`
 			Sha256   string `json:"sha256"    validate:"required"`
 		} `json:"audio"`
+		Button *struct {
+			Payload string `json:"payload"`
+			Text    string `json:"text"    validate:"required"`
+		} `json:"button"`
 		Document *struct {
 			File     string `json:"file"      validate:"required"`
 			ID       string `json:"id"        validate:"required"`
@@ -110,6 +114,7 @@ type eventPayload struct {
 			MimeType string `json:"mime_type" validate:"required"`
 			Sha256   string `json:"sha256"    validate:"required"`
 			Caption  string `json:"caption"`
+			Filename string `json:"filename"`
 		} `json:"document"`
 		Image *struct {
 			File     string `json:"file"      validate:"required"`
@@ -190,6 +195,8 @@ func (h *handler) receiveEvent(ctx context.Context, channel courier.Channel, w h
 			text = msg.Text.Body
 		} else if msg.Type == "audio" && msg.Audio != nil {
 			mediaURL, err = resolveMediaURL(channel, msg.Audio.ID)
+		} else if msg.Type == "button" && msg.Button != nil {
+			text = msg.Button.Text
 		} else if msg.Type == "document" && msg.Document != nil {
 			text = msg.Document.Caption
 			mediaURL, err = resolveMediaURL(channel, msg.Document.ID)
@@ -336,9 +343,10 @@ type mtTextPayload struct {
 }
 
 type mediaObject struct {
-	ID      string `json:"id,omitempty"`
-	Link    string `json:"link,omitempty"`
-	Caption string `json:"caption,omitempty"`
+	ID       string `json:"id,omitempty"`
+	Link     string `json:"link,omitempty"`
+	Caption  string `json:"caption,omitempty"`
+	Filename string `json:"filename,omitempty"`
 }
 
 type LocalizableParam struct {
@@ -467,6 +475,12 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				if attachmentCount == 0 {
 					mediaPayload.Caption = msg.Text()
 				}
+				mediaPayload.Filename, err = utils.BasePathForURL(mediaURL)
+
+				// Logging error
+				if err != nil {
+					logrus.WithField("channel_uuid", msg.Channel().UUID().String()).WithError(err).Error("Error while parsing the media URL")
+				}
 				payload.Document = mediaPayload
 				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			} else if strings.HasPrefix(mimeType, "image") {
@@ -525,7 +539,10 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 		}
 
 		if templating != nil {
-			namespace := msg.Channel().StringConfigForKey(configNamespace, "")
+			namespace := templating.Namespace
+			if namespace == "" {
+				namespace = msg.Channel().StringConfigForKey(configNamespace, "")
+			}
 			if namespace == "" {
 				return nil, errors.Errorf("cannot send template message without Facebook namespace for channel: %s", msg.Channel().UUID())
 			}
@@ -915,6 +932,7 @@ type MsgTemplating struct {
 	} `json:"template" validate:"required,dive"`
 	Language  string   `json:"language" validate:"required"`
 	Country   string   `json:"country"`
+	Namespace string   `json:"namespace"`
 	Variables []string `json:"variables"`
 }