+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+v6.3.5
+----------
+ * up max request size to 1M
+
+v6.3.4
+----------
+ * Include filename when sending WhatsApp attachments
+
+v6.3.3
+----------
+ * Support using namespace from the template translation
+ * Add is_resend to Msg payload to allow for resending messages manually
+
+v6.3.2
+----------
+ * Do not verify the SSL certificate for Bongo Live
+
+v6.3.1
+----------
+ * Update BL to remove UDH parameter and use HTTPS URL
+
+v6.2.2
+----------
+ * Handle whatsapp URNs sent to Twiml handler without prefix
+ * Add support for Zenvia SMS
+
+v6.2.1
+----------
+ * Add support for Zenvia WhatsApp
+
+v6.2.0
+----------
+ * Add handling for button whatsapp message type
+ * Bump CI testing to PG 11 and 12
+ * Add Kaleyra channel type
+ * 6.2.0 RC
+
+v6.1.7
+----------
+ * switch id to bigserial
+
 v6.1.6
 ----------
  * Cache media upload failures localy for 15m